+++ ledger/src/blockstore_meta.rs
@@ -469,7 +469,7 @@ impl ShredIndex {
         }
     }
 
-    pub(crate) fn contains(&self, idx: u64) -> bool {
+    pub fn contains(&self, idx: u64) -> bool {
         self.index.contains(idx as usize)
     }