+++ net-utils/src/lib.rs
@@ -64,7 +64,7 @@ pub const VALIDATOR_PORT_RANGE: PortRange = (
     crate::sockets::UNIQUE_ALLOC_BASE_PORT,
 );
 
-pub const MINIMUM_VALIDATOR_PORT_RANGE_WIDTH: u16 = 25; // VALIDATOR_PORT_RANGE must be at least this wide
+pub const MINIMUM_VALIDATOR_PORT_RANGE_WIDTH: u16 = 26; // VALIDATOR_PORT_RANGE must be at least this wide
 
 pub(crate) const HEADER_LENGTH: usize = 4;
 pub(crate) const IP_ECHO_SERVER_RESPONSE_LENGTH: usize = HEADER_LENGTH + 23;