+++ votor/src/event.rs
@@ -2,6 +2,7 @@ use {
     agave_votor_messages::consensus_message::Block,
     crossbeam_channel::{Receiver, Sender},
     solana_clock::Slot,
+    solana_hash::Hash,
     solana_runtime::bank::Bank,
     std::{sync::Arc, time::Instant},
 };
@@ -93,6 +94,27 @@ impl VotorEvent {
             VotorEvent::Standstill(s) => s < &root,
             VotorEvent::ProduceWindow(info) => info.start_slot <= root,
             VotorEvent::SetIdentity => false,
+        }
+    }
+}
+
+pub type RepairEventSender = Sender<RepairEvent>;
+pub type RepairEventReceiver = Receiver<RepairEvent>;
+
+/// Event sent by votor to the block id repair service for informed repair
+#[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
+pub enum RepairEvent {
+    /// We require that this block be fetched. This can happen for the following reasons:
+    /// - The block has received a NotarizeFallback certificate or stronger
+    /// - TODO(ashwin): An intrawindow block has reached the SafeToNotar threshold, however we need
+    ///   to check that the parent has reached notarize-fallback requiring us to fetch this block
+    FetchBlock { slot: Slot, block_id: Hash },
+}
+
+impl RepairEvent {
+    pub fn slot(&self) -> Slot {
+        match self {
+            RepairEvent::FetchBlock { slot, .. } => *slot,
         }
     }
 }